only int

+++ src/main/java/Board.java
@@ -6,33 +6,33 @@ import edu.princeton.cs.algs4.StdOut;
 
 public class Board {
     private ArrayList<Board> neighbors;
-    private final short manhattanSum;
+    private final int manhattanSum;
     private final int hammingCount;
     private final int dimension;
     private final int key;
     private int zeroIndex;
     private Board twinBoard;
-    private final short[] blocks;
+    private final int[] blocks;
     // construct a board from an n-by-n array of blocks
     public Board(int[][] arr) {
         dimension = arr.length;
-        blocks = new short[dimension*dimension];
+        blocks = new int[dimension*dimension];
 
-        short k = 0;
+        int k = 0;
         for (int i = 0; i < dimension; i++) {
             for (int j = 0; j < dimension; j++) {
                 if (arr[i][j] == 0) {
                     zeroIndex = k;
                 }
-                blocks[k++] = (short) arr[i][j];
+                blocks[k++] = (int) arr[i][j];
             }
         }
 
-        short[] copy = Arrays.copyOf(blocks, blocks.length);
+        int[] copy = Arrays.copyOf(blocks, blocks.length);
         exch(copy, zeroIndex, copy.length - 1);
         Arrays.sort(copy, 0, blocks.length - 1);
 
-        short sum = 0;
+        int sum = 0;
         int hc = 0;
         for (int i = 0; i < copy.length; i++) {
             int currentValue = blocks[i];
@@ -47,7 +47,7 @@ public class Board {
         key = blocks.hashCode();
     }
 
-    private static int[][] getMatrix(short[] arr, int n) {
+    private static int[][] getMatrix(int[] arr, int n) {
         int[][] m = new int[n][n];
         int k = 0;
         for (int i = 0; i < n; i++) {
@@ -58,7 +58,7 @@ public class Board {
         return m;
     }
 
-    private static int biSearch(short[] arr, int from, int to, int targetValue) {
+    private static int biSearch(int[] arr, int from, int to, int targetValue) {
         int mid;
         do {
             mid = (from + to) >>> 1;
@@ -74,8 +74,8 @@ public class Board {
 
     private static int[] getMatrixCoordsByIndex(int index, int n) {
         int[] coords = {
-            (short) (index / n),
-            (short) (index % n)
+            (int) (index / n),
+            (int) (index % n)
         };
         return coords;
     }
@@ -90,11 +90,11 @@ public class Board {
         return hammingCount;
     }
 
-    private short getMhDistance(int[] c1, int[] c2) {
-        return (short) (Math.abs(c1[0] - c2[0]) + Math.abs(c2[1] - c1[1]));
+    private int getMhDistance(int[] c1, int[] c2) {
+        return (int) (Math.abs(c1[0] - c2[0]) + Math.abs(c2[1] - c1[1]));
     }
 
-    private short getManhattanDistance(int fromIndex, int toIndex, int size) {
+    private int getManhattanDistance(int fromIndex, int toIndex, int size) {
         int[] from = getMatrixCoordsByIndex(fromIndex, size);
         int[] to = getMatrixCoordsByIndex(toIndex, size);
         return getMhDistance(
@@ -121,7 +121,7 @@ public class Board {
         return twinBoard;
     }
 
-    private static Board getTwinBoard(int zeroIndex, short[] m, int size) {
+    private static Board getTwinBoard(int zeroIndex, int[] m, int size) {
         int r1 = -1;
         int r2 = -1;
         for (int i = 0; i < size*size; i++) {
@@ -157,13 +157,13 @@ public class Board {
         return neighbors;
     }
 
-    private static void exch(short[] arr, int x, int y) {
-        short v = arr[x];
+    private static void exch(int[] arr, int x, int y) {
+        int v = arr[x];
         arr[x] = arr[y];
         arr[y] = v;
     }
 
-    private static ArrayList<Board> findNeighbors(int zeroIndex, short[] blocks, int n) {
+    private static ArrayList<Board> findNeighbors(int zeroIndex, int[] blocks, int n) {
         int[] zeroCoords = getMatrixCoordsByIndex(zeroIndex, n);
         int x = zeroCoords[0];
         int y = zeroCoords[1];
